MFC: r215093
Enable reservation-based physical memory allocation.  Even without the
creation of large page mappings in the pmap, it can provide modest
performance benefits.  In particular, for a "buildworld" on a 2x 1GHz
Ultrasparc IIIi it reduced the wall clock time by 2.2% and the system
time by 12.6%.

Tested by:	marius@

+++ sys/vm/vm_reserv.c
@@ -197,7 +197,7 @@ sysctl_vm_reserv_partpopq(SYSCTL_HANDLER_ARGS)
 		}
 		mtx_unlock(&vm_page_queue_free_mtx);
 		sbuf_printf(&sbuf, "%5d: %6dK, %6d\n", level,
-		    unused_pages * (PAGE_SIZE / 1024), counter);
+		    unused_pages * ((int)PAGE_SIZE / 1024), counter);
 	}
 	sbuf_finish(&sbuf);
 	error = SYSCTL_OUT(req, sbuf_data(&sbuf), sbuf_len(&sbuf));
